backend/requirements.txt
fastapi
uvicorn
pydantic
requests
boto3
botocore
mangum
load_dotenv
tiktoken